// frontend/src/api/generated/model/createPostRequestType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type CreatePostRequestType = typeof CreatePostRequestType[keyof typeof CreatePostRequestType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreatePostRequestType = {
  LINK: 'LINK',
  ASK: 'ASK',
} as const;